各レスポンスを作成

+++ src/main/java/com/example/demo/controller/AnimeController.java
@@ -3,6 +3,10 @@ package com.example.demo.controller;
 import com.example.demo.entity.Anime;
 import com.example.demo.form.AnimeForm;
 import com.example.demo.service.AnimeService;
+import com.example.demo.service.ResponseService;
+import com.example.demo.service.response.Responce;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,9 +17,11 @@ import java.util.Optional;
 public class AnimeController {
 
     private final AnimeService animeService;
+    private final ResponseService responseService;
 
-    public AnimeController(AnimeService animeService) {
+    public AnimeController(AnimeService animeService, ResponseService responseService) {
         this.animeService = animeService;
+        this.responseService = responseService;
     }
 
     @GetMapping
@@ -29,17 +35,34 @@ public class AnimeController {
     }
 
     @PostMapping
-    public void registerAnime(@RequestBody AnimeForm animeForm) {
-        animeService.registerAnime(animeForm.getName(), animeForm.getGenre());
+    public Responce registerAnime(@Validated @RequestBody AnimeForm animeForm, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return responseService.validationError(bindingResult);
+        } else {
+            animeService.registerAnime(animeForm.getName(), animeForm.getGenre());
+            return responseService.postSuccess();
+        }
     }
 
     @PatchMapping(path = "{animeId}")
-    public void updateAnime(@PathVariable("animeId") Integer id, @RequestBody AnimeForm animeForm) {
-        animeService.updateAnime(id, animeForm.getName(), animeForm.getGenre());
+    public Responce updateAnime(@PathVariable("animeId") Integer id, @Validated @RequestBody AnimeForm animeForm, BindingResult bindingResult) {
+        if (animeService.getAnime(id).isEmpty()) {
+            return responseService.userNotFound();
+        } else if (bindingResult.hasErrors()) {
+            return responseService.validationError(bindingResult);
+        } else {
+            animeService.updateAnime(id, animeForm.getName(), animeForm.getGenre());
+            return responseService.updateSuccess();
+        }
     }
 
+
     @DeleteMapping(path = "{animeId}")
-    public void deleteAnime(@PathVariable("animeId") Integer id) {
+    public Responce deleteAnime(@PathVariable("animeId") Integer id) {
+        if (animeService.getAnime(id).isEmpty()) {
+            return responseService.userNotFound();
+        }
         animeService.deleteAnime(id);
+        return responseService.deleteSuccess();
     }
 }
